agents/remote/kafka-user/requirements.txt
google-adk[a2a]
google-genai
python-dotenv
httpx